+++ l10n_br_account_product/models/l10n_br_account_product.py
@@ -26,32 +26,6 @@ NFE_IND_IE_DEST = [
 NFE_IND_IE_DEST_DEFAULT = NFE_IND_IE_DEST[0][0]
 
 
-class L10nBrAccountServiceType(models.Model):
-    _name = 'l10n_br_account.service.type'
-    _description = u'Cadastro de Operações Fiscais de Serviço'
-
-    code = fields.Char(u'Código', size=16, required=True)
-    name = fields.Char(u'Descrição', size=256, required=True)
-    parent_id = fields.Many2one(
-        'l10n_br_account.service.type', 'Tipo de Serviço Pai')
-    child_ids = fields.One2many(
-        'l10n_br_account.service.type', 'parent_id',
-        u'Tipo de Serviço Filhos')
-    internal_type = fields.Selection(
-        [('view', u'Visualização'), ('normal', 'Normal')], 'Tipo Interno',
-        required=True, default='normal')
-
-    @api.multi
-    def name_get(self):
-        result = []
-        for record in self:
-            name = record['name']
-            if record['code']:
-                name = record['code'] + ' - ' + name
-            result.append((record['id'], name))
-        return result
-
-
 class L10nbrAccountDocumentRelated(models.Model):
     _name = 'l10n_br_account_product.document.related'
