+++ l10n_br_account_product/models/l10n_br_account_service_type.py
@@ -1,0 +1,32 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2013  Renato Lima - Akretion
+# License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
+
+from openerp import models, fields, api
+
+
+class L10nBrAccountServiceType(models.Model):
+    _name = 'l10n_br_account.service.type'
+    _description = u'Cadastro de Operações Fiscais de Serviço'
+
+    code = fields.Char(u'Código', size=16, required=True)
+    name = fields.Char(u'Descrição', size=256, required=True)
+    parent_id = fields.Many2one(
+        'l10n_br_account.service.type', 'Tipo de Serviço Pai')
+    child_ids = fields.One2many(
+        'l10n_br_account.service.type', 'parent_id',
+        u'Tipo de Serviço Filhos')
+    internal_type = fields.Selection(
+        [('view', u'Visualização'), ('normal', 'Normal')], 'Tipo Interno',
+        required=True, default='normal')
+
+    @api.multi
+    def name_get(self):
+        result = []
+        for record in self:
+            name = record['name']
+            if record['code']:
+                name = record['code'] + ' - ' + name
+            result.append((record['id'], name))
+        return result
+
